genie-skill-bot/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
azure-core==1.32.0
botbuilder-core==4.16.2
botbuilder-integration-aiohttp==4.16.2
botbuilder-schema==4.16.2
botframework-connector==4.16.2
botframework-streaming==4.16.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
frozenlist==1.5.0
idna==3.10
isodate==0.7.2
jsonpickle==1.4.2
msal==1.31.0
msrest==0.7.1
multidict==6.1.0
oauthlib==3.2.2
propcache==0.2.0
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.17.1
